+++ src/hapi/pipelines/app/pipelines.py
@@ -31,6 +31,7 @@ from hapi.pipelines.database.refugees_and_returnees import RefugeesAndReturnees
 from hapi.pipelines.database.sector import Sector
 from hapi.pipelines.database.wfp_commodity import WFPCommodity
 from hapi.pipelines.database.wfp_market import WFPMarket
+from hapi.pipelines.utilities.error_handling import ErrorManager
 
 logger = logging.getLogger(__name__)
 
@@ -57,6 +58,7 @@ class Pipelines:
             countries=countries_to_run,
         )
         self.countries = self.locations.hapi_countries
+        self.error_manager = ErrorManager()
         reader = Read.get_reader("hdx")
         libhxl_dataset = AdminLevel.get_libhxl_dataset(
             retriever=reader
@@ -171,13 +173,6 @@ class Pipelines:
                 current_scrapers + scraper_names
             )
 
-        _create_configurable_scrapers("population", "national")
-        _create_configurable_scrapers(
-            "population", "adminone", adminlevel=self.adminone
-        )
-        _create_configurable_scrapers(
-            "population", "admintwo", adminlevel=self.admintwo
-        )
         _create_configurable_scrapers(
             "operational_presence", "admintwo", adminlevel=self.admintwo
         )
@@ -194,7 +189,6 @@ class Pipelines:
         _create_configurable_scrapers(
             "idps", "admintwo", adminlevel=self.admintwo
         )
-        _create_configurable_scrapers("poverty_rate", "national")
         _create_configurable_scrapers("conflict_event", "national")
         _create_configurable_scrapers(
             "conflict_event", "admintwo", adminlevel=self.admintwo
@@ -205,14 +199,12 @@ class Pipelines:
 
     def output_population(self):
         if not self.themes_to_run or "population" in self.themes_to_run:
-            results = self.runner.get_hapi_results(
-                self.configurable_scrapers["population"]
-            )
             population = Population(
                 session=self.session,
                 metadata=self.metadata,
                 admins=self.admins,
-                results=results,
+                configuration=self.configuration,
+                error_manager=self.error_manager,
             )
             population.populate()
 
@@ -235,6 +227,7 @@ class Pipelines:
                 sector=self.sector,
                 results=results,
                 config=self.configuration,
+                error_manager=self.error_manager,
             )
             operational_presence.populate()
 
@@ -248,6 +241,7 @@ class Pipelines:
                 admintwo=self.admintwo,
                 countryiso3s=self.countries,
                 configuration=self.configuration,
+                error_manager=self.error_manager,
             )
             food_security.populate()
 
@@ -262,6 +256,7 @@ class Pipelines:
                 admins=self.admins,
                 sector=self.sector,
                 configuration=self.configuration,
+                error_manager=self.error_manager,
             )
             humanitarian_needs.populate()
 
@@ -304,6 +299,7 @@ class Pipelines:
                 metadata=self.metadata,
                 admins=self.admins,
                 results=results,
+                error_manager=self.error_manager,
             )
             idps.populate()
 
@@ -315,20 +311,18 @@ class Pipelines:
                 countryiso3s=self.countries,
                 locations=self.locations,
                 configuration=self.configuration,
+                error_manager=self.error_manager,
             )
             funding.populate()
 
     def output_poverty_rate(self):
         if not self.themes_to_run or "poverty_rate" in self.themes_to_run:
-            results = self.runner.get_hapi_results(
-                self.configurable_scrapers["poverty_rate"]
-            )
             poverty_rate = PovertyRate(
                 session=self.session,
                 metadata=self.metadata,
                 admins=self.admins,
-                config=self.configuration["poverty_rate_national"],
-                results=results,
+                configuration=self.configuration,
+                error_manager=self.error_manager,
             )
             poverty_rate.populate()
 
@@ -342,7 +336,8 @@ class Pipelines:
                 metadata=self.metadata,
                 admins=self.admins,
                 results=results,
-                config=self.configuration,
+                configuration=self.configuration,
+                error_manager=self.error_manager,
             )
             conflict_event.populate()
 
@@ -361,6 +356,7 @@ class Pipelines:
                 adminone=self.adminone,
                 admintwo=self.admintwo,
                 configuration=self.configuration,
+                error_manager=self.error_manager,
             )
             wfp_market.populate()
             food_price = FoodPrice(
@@ -371,6 +367,7 @@ class Pipelines:
                 currency=self.currency,
                 commodity=wfp_commodity,
                 market=wfp_market,
+                error_manager=self.error_manager,
             )
             food_price.populate()
 
@@ -396,3 +393,6 @@ class Pipelines:
 
     def debug(self, folder: str) -> None:
         self.org.output_org_map(folder)
+
+    def output_errors(self, err_to_hdx: bool) -> None:
+        self.error_manager.output_errors(err_to_hdx)